// app/types.ts
export interface IPosition {
	name: string;
	percent: number;
}

export interface IRank {
	name: string;
	summa: number;
}

export interface IRegion {
	name: string;
	type: string;
}

export interface IEmployee {
	FIOlotin: string;
	FIOkril: string;
	inn: number;
	inps: number;
	plastic: number;
	dateOfEmployment: string;
}